Fixed bug in old assign_dimensions, now called deint_z. Added blur_otsu_contours function.
The old assign_dimensions function miscategorized the z and t dimensions, resulting in impoper ordering of the ndarry. The bug was fixed. Also added blur_otsu_contours function that finds the contours of an image after applying median blurring and thresholding. The contours_beta.py is the script I used to validate the blur_otsu_contours function.